bug 1229. FIxed check to see if file was sucessfully renamed
git-svn-id: https://svn.broadinstitute.org/genepattern/trunk@2916 48a7a3f7-ad46-0410-b969-29950918272e

+++ website/saveTask.jsp
@@ -350,11 +350,14 @@ try {
 			attachment = new File(dir, attachmentName);
 			if (attachment.exists()) {
 				File oldVersion = new File(dir, attachment.getName() + ".old");
-				oldVersion.delete(); // delete the previous .old file
-				boolean renamed = attachment.renameTo(oldVersion);
-				Vector v = new Vector();
-				v.add("failed to rename " + oldVersion.getName() +".");
-				throw new WebServiceErrorMessageException(v);
+				if(oldVersion.exists()) {
+					oldVersion.delete(); // delete the previous .old file
+				}
+				if(!attachment.renameTo(oldVersion)) {
+					Vector v = new Vector();
+					v.add("failed to rename " + oldVersion.getName() +".");
+					throw new WebServiceErrorMessageException(v);
+				}
 			}
 			
 			attachedFile.saveAs(dir.getPath() + File.separator + attachmentName);
